+++ R/fia_tidy.R
@@ -14,7 +14,7 @@
 #'
 #' @param db a list of tables produced by [fia_load()]
 #' @export
-#' @seealso [add_composite_ids()]
+#' @seealso [fia_add_composite_ids()]
 #' @returns a tibble
 fia_tidy <- function(db) {
   # Select only the columns we need from each table, to keep things slim
@@ -29,7 +29,7 @@ fia_tidy <- function(db) {
     db$PLOT |>
     dplyr::filter(INVYR >= 2000L) |>
     dplyr::mutate(CN = as.character(CN)) |> 
-    add_composite_ids() |>
+    fia_add_composite_ids() |>
     dplyr::select(
       plot_ID,
       PLT_CN = CN,
@@ -43,7 +43,7 @@ fia_tidy <- function(db) {
     db$COND |>
     dplyr::filter(INVYR >= 2000L) |>
     dplyr::mutate(PLT_CN = as.character(PLT_CN)) |> 
-    add_composite_ids() |>
+    fia_add_composite_ids() |>
     dplyr::select(
       plot_ID,
       PLT_CN,
@@ -59,7 +59,7 @@ fia_tidy <- function(db) {
     db$TREE |>
     dplyr::filter(INVYR >= 2000L) |>
     dplyr::mutate(PLT_CN = as.character(PLT_CN)) |> 
-    add_composite_ids() |>
+    fia_add_composite_ids() |>
     dplyr::select(
       plot_ID,
       tree_ID,
@@ -139,84 +139,4 @@ fia_tidy <- function(db) {
 
   # return:
   data
-}
-
-#' Add composite ID columns to data
-#'
-#' Creates a `tree_ID` and/or a `plot_ID` column that contain unique tree and
-#' plot identifiers, respectively.  These are created by pasting together the
-#' values for `STATECD`, `UNITCD`, `COUNTYCD`, `PLOT` and in the case of trees
-#' `SUBP` and `TREE`.
-#'
-#' @param data A tibble or data frame with at least the `STATECD`, `UNITCD`,
-#'   `COUNTYCD` and `PLOT` columns
-#'
-#' @seealso See [split_composite_ids()] for "undoing" this.
-#' @returns The input tibble with a `plot_ID` and possibly also a `tree_ID`
-#'   column added
-#' @export
-add_composite_ids <- function(data) {
-  cols <- colnames(data)
-  if (
-    all(c("STATECD", "UNITCD", "COUNTYCD", "PLOT", "SUBP", "TREE") %in% cols)
-  ) {
-    data <-
-      data |>
-      dplyr::mutate(
-        plot_ID = paste(STATECD, UNITCD, COUNTYCD, PLOT, sep = "_"),
-        tree_ID = paste(STATECD, UNITCD, COUNTYCD, PLOT, SUBP, TREE, sep = "_"),
-        .before = 1
-      )
-  } else if (all(c("STATECD", "UNITCD", "COUNTYCD", "PLOT") %in% cols)) {
-    data <-
-      data |>
-      dplyr::mutate(
-        plot_ID = paste(STATECD, UNITCD, COUNTYCD, PLOT, sep = "_"),
-        .before = 1
-      )
-  } else {
-    stop("Not all required columns are present")
-  }
-  data
-}
-
-#' Split composite ID columns
-#'
-#' Splits the composite ID columns `tree_ID` and/or `plot_ID` into their
-#' original component columns
-#'
-#' @param data A tibble with the `tree_ID` and/or `plot_ID` columns
-#' @returns The input tibble with additional columns `STATECD`, `UNITCD`,
-#'   `COUNTYCD`, `PLOT` and possibly `SUBP` and `TREE`.
-#' @seealso [add_composite_ids()]
-#' @export
-split_composite_ids <- function(data) {
-  cols <- colnames(data)
-  if (!any(c("plot_ID", "tree_ID") %in% cols)) {
-    stop("No composite ID columns found")
-  }
-
-  # tree_ID contains all the information in plot_ID, so if tree_ID exists, it's
-  # enough to just split that one
-  if ("tree_ID" %in% cols) {
-    data <- data |>
-      tidyr::separate_wider_delim(
-        tree_ID,
-        delim = "_",
-        names = c("STATECD", "UNITCD", "COUNTYCD", "PLOT", "SUBP", "TREE"),
-        cols_remove = FALSE
-      )
-    return(data)
-  }
-
-  if ("plot_ID" %in% cols) {
-    data <- data |>
-      tidyr::separate_wider_delim(
-        plot_ID,
-        delim = "_",
-        names = c("STATECD", "UNITCD", "COUNTYCD", "PLOT"),
-        cols_remove = FALSE
-      )
-    return(data)
-  }
 }